+++ src/app/(general)/layout.tsx
@@ -1,5 +1,6 @@
 import { SidebarProvider } from "@/components/shadcnui/sidebar";
 import Header from "@/components/shared/Header";
+import { AppSidebar } from "@/components/shared/Sidebar";
 
 export default function GeneralLayout({
   children,
@@ -10,6 +11,8 @@ export default function GeneralLayout({
     <SidebarProvider defaultOpen={false}>
       <div className="w-full">
         <Header />
+        <AppSidebar />
+
         {children}
       </div>
     </SidebarProvider>
