+++ src/components/shared/Sidebar/index.tsx
@@ -2,7 +2,6 @@
 
 import {
   LayoutDashboard,
-  ChevronLeft,
   ChevronRight,
   Landmark,
   CreditCard,
@@ -21,7 +20,6 @@ import {
 } from "@/components/shadcnui/sidebar";
 import Link from "next/link";
 import Image from "next/image";
-import { usePathname } from "next/navigation";
 
 const items = [
   { title: "Dashboard", url: "/dashboard", icon: LayoutDashboard },
@@ -32,41 +30,13 @@ const items = [
 ];
 
 export function AppSidebar() {
-  const pathname = usePathname();
-  const { open, openMobile, toggleSidebar } = useSidebar();
-
-  const currentItem = items.find((item) => pathname.startsWith(item.url));
-  const currentTitle = currentItem ? currentItem.title : "Dashboard";
+  const { toggleSidebar } = useSidebar();
 
   return (
-    <div className="flex">
-      {/* open on mobile */}
-      <button
-        onClick={() => toggleSidebar()}
-        className="flex md:hidden items-center z-50 p-2 text-xl font-bold text-primary"
-      >
-        {openMobile ? (
-          <ChevronRight size={23} strokeWidth={3} />
-        ) : (
-          <ChevronLeft size={23} strokeWidth={3} />
-        )}
-        <span>{currentTitle}</span>
-      </button>
+    <div className="flex relative">
 
-      {/* open on desktop */}
-      <button
-        onClick={() => toggleSidebar()}
-        className="hidden md:flex items-center z-50 p-2 text-xl font-bold text-primary"
-      >
-        {!open && (
-          <>
-            <ChevronLeft size={23} strokeWidth={3} />
-            <span>Dashboard</span>
-          </>
-        )}
-      </button>
-
-      <Sidebar variant="floating" side="right" collapsible="offcanvas">
+      {/** Sidebar só exibe se for mobile OU desktop com `open` */}
+      <Sidebar side="right" collapsible="offcanvas">
         <SidebarHeader className="pt-4">
           <button
             onClick={() => toggleSidebar()}
@@ -86,7 +56,11 @@ export function AppSidebar() {
             {items.map((item) => (
               <SidebarMenuItem key={item.title}>
                 <SidebarMenuButton asChild>
-                  <Link onClick={toggleSidebar} href={item.url} className="flex items-center gap-2">
+                  <Link
+                    onClick={toggleSidebar}
+                    href={item.url}
+                    className="flex items-center gap-2"
+                  >
                     <span className="text-2xl font-bold hover:underline group-hover:text-primary">
                       {item.title}
                     </span>
